+++ OfficeManagerMini/02_Services/OMM.Services.YWeather/YWeather/Forecast.cs
@@ -3,7 +3,7 @@
     public class Forecast
     {
         public string Day { get; set; }
-        public int Date { get; set; }
+        public long Date { get; set; }
         public int Low { get; set; }
         public int High { get; set; }
         public string Text { get; set; }
